Created test groups, so that all integration tests run when the docker image is up
