+++ src/test/java/com/flagsmith/FlagsmithClientHttpErrorsTest.java
@@ -32,7 +32,7 @@ public class FlagsmithClientHttpErrorsTest {
                 .build();
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_Features_Then_Empty() {
         List<Flag> featureFlags = flagsmithClient.getFeatureFlags();
 
@@ -40,7 +40,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertTrue(featureFlags.isEmpty(), "Should not have test featureFlags back");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_Features_For_User_Then_Empty() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -52,7 +52,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertTrue(featureFlags.isEmpty(), "Should not have test featureFlags back");
     }
 
-    @Test(groups = "integration", expectedExceptions = FlagsmithException.class)
+    @Test(groups = "integration-offline", expectedExceptions = FlagsmithException.class)
     public void testClient_When_Get_Features_For_User_Then_Throw() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -61,7 +61,7 @@ public class FlagsmithClientHttpErrorsTest {
         flagsmithClient.getFeatureFlags(user, true);
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Traits_Then_Null() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -72,7 +72,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(userTraits, "Should have user traits back");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Traits_For_Keys_Then_Null() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -83,7 +83,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(userTraits, "Should have user traits back");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Traits_For_Invalid_User_Then_Return_Null() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -94,7 +94,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(userTraits, "Should have user traits back");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Traits_And_Flags_For_Keys_Then_Null_Lists() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -107,7 +107,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(userFlagsAndTraits.getTraits(), "Should not have user traits back");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Traits_And_Flags_For_Invalid_User_Then_Return_Null_Lists() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -120,7 +120,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(userFlagsAndTraits.getTraits(), "Should not have no user traits back");
     }
 
-    @Test(groups = "integration", expectedExceptions = FlagsmithException.class)
+    @Test(groups = "integration-offline", expectedExceptions = FlagsmithException.class)
     public void testClient_When_Get_User_Traits_And_Flags_Then_Throw() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -129,7 +129,7 @@ public class FlagsmithClientHttpErrorsTest {
         flagsmithClient.getUserFlagsAndTraits(user, true);
     }
 
-    @Test(groups = "integration", expectedExceptions = FlagsmithException.class)
+    @Test(groups = "integration-offline", expectedExceptions = FlagsmithException.class)
     public void testClient_When_Get_User_Traits_And_Flags_Then_Throw_evenIfLoggingDisabled() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -144,7 +144,7 @@ public class FlagsmithClientHttpErrorsTest {
         flagsmithClient.getUserFlagsAndTraits(user, true);
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Trait_From_Traits_And_Flags_For_Keys_Then_Null() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -157,7 +157,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(userTrait, "Should not have user traits back");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Traits_From_Traits_And_Flags_For_Keys_Then_Null() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -170,7 +170,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(traits, "Should not have user traits back");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_FLag_Value_From_Traits_And_Flags_For_Keys_Then_Null() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -183,7 +183,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertNull(featureFlagValue, "Should not have feature");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_FLag_Enabled_From_Traits_And_Flags_For_Keys_Then_False() {
         // context user
         FeatureUser user = new FeatureUser();
@@ -196,7 +196,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertFalse(enabled, "Should not have feature enabled");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Trait_Then_Null() {
         FeatureUser user = new FeatureUser();
         user.setIdentifier("another_user");
@@ -207,7 +207,7 @@ public class FlagsmithClientHttpErrorsTest {
     }
 
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Get_User_Trait_Update_Then_Null() {
         FeatureUser user = new FeatureUser();
         user.setIdentifier("another_user");
@@ -221,7 +221,7 @@ public class FlagsmithClientHttpErrorsTest {
     }
 
 
-    @Test(groups = "integration", expectedExceptions = FlagsmithException.class)
+    @Test(groups = "integration-offline", expectedExceptions = FlagsmithException.class)
     public void testClient_When_Get_User_Trait_Update_Then_Throw() {
         FeatureUser user = new FeatureUser();
         user.setIdentifier("another_user");
@@ -233,7 +233,7 @@ public class FlagsmithClientHttpErrorsTest {
         flagsmithClient.updateTrait(user, userTrait, true);
     }
 
-    @Test(expectedExceptions = IllegalArgumentException.class, groups = "integration")
+    @Test(expectedExceptions = IllegalArgumentException.class, groups = "integration-offline")
     public void testClient_When_Add_Traits_For_Identity_With_Missing_Identity_Then_Failed() {
         // Given traits and no user Identity
         Trait trait1 = new Trait();
@@ -248,7 +248,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertTrue(traits.size() == 0, "Should not return any traits");
     }
 
-    @Test(groups = "integration")
+    @Test(groups = "integration-offline")
     public void testClient_When_Add_Traits_For_Identity_Then_Success() {
         // Given
         FeatureUser user = new FeatureUser();
@@ -269,7 +269,7 @@ public class FlagsmithClientHttpErrorsTest {
         assertTrue(traits.isEmpty(), "Should not have traits returned");
     }
 
-    @Test(groups = "integration", expectedExceptions = FlagsmithException.class)
+    @Test(groups = "integration-offline", expectedExceptions = FlagsmithException.class)
     public void testClient_When_Add_Traits_For_Identity_Then_Throw() {
         // Given
         FeatureUser user = new FeatureUser();
