+++ src/test/java/com/flagsmith/FlagsmithClientTest.java
@@ -1,27 +1,21 @@
 package com.flagsmith;
 
-import com.google.common.collect.ImmutableMap;
-import io.restassured.RestAssured;
-import io.restassured.http.Header;
-import io.restassured.http.Headers;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.containers.Network;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.containers.output.Slf4jLogConsumer;
-import org.testcontainers.containers.wait.strategy.HttpWaitStrategy;
-import org.testcontainers.utility.DockerImageName;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
+import static com.flagsmith.FlagsmithTestHelper.assignTraitToUserIdentity;
+import static com.flagsmith.FlagsmithTestHelper.config;
+import static com.flagsmith.FlagsmithTestHelper.createFeature;
+import static com.flagsmith.FlagsmithTestHelper.createProjectEnvironment;
+import static com.flagsmith.FlagsmithTestHelper.createUserIdentity;
+import static com.flagsmith.FlagsmithTestHelper.featureUser;
+import static com.flagsmith.FlagsmithTestHelper.flag;
+import static com.flagsmith.FlagsmithTestHelper.switchFlag;
+import static com.flagsmith.FlagsmithTestHelper.switchFlagForUser;
+import static com.flagsmith.FlagsmithTestHelper.trait;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -31,83 +25,29 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
  */
 @Test(groups = "integration")
 public class FlagsmithClientTest {
-    private static final Logger log = LoggerFactory.getLogger(FlagsmithClientTest.class);
-
-    private static final Network network = Network.newNetwork();
-    protected static final int BACKEND_PORT = 8000;
-
-    private static PostgreSQLContainer<?> postgres;
-
-    private static GenericContainer<?> backend;
-
-    private static String token;
-
-    private static int organisationId;
-
-    @BeforeClass
-    public static void beforeClass() {
-        postgres = new PostgreSQLContainer<>(DockerImageName.parse("postgres:10.6-alpine"))
-                .withNetwork(network)
-                .withNetworkAliases("flagsmith-db")
-                .withDatabaseName("flagsmith")
-                .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("flagsmith-db")));
-        postgres.start();
-
-        backend = new GenericContainer<>(DockerImageName.parse("flagsmith/flagsmith-api:v2-5-2"))
-                .withNetwork(network)
-                .withNetworkAliases("flagsmith-be")
-                .withEnv("DJANGO_ALLOWED_HOSTS", "*")
-                .withEnv("DATABASE_URL", String.format("postgresql://%s:%s@%s:%d/%s",
-                        postgres.getUsername(),
-                        postgres.getPassword(),
-                        "flagsmith-db",
-                        PostgreSQLContainer.POSTGRESQL_PORT,
-                        postgres.getDatabaseName()))
-                .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger("flagsmith-be")))
-                .withExposedPorts(BACKEND_PORT)
-                .waitingFor(new HttpWaitStrategy().forPath("/health").withHeader("Accept", "application/json"));
-        backend.start();
-
-        RestAssured.enableLoggingOfRequestAndResponseIfValidationFails();
-        RestAssured.port = backend.getMappedPort(BACKEND_PORT);
-
-        assertThat(RestAssured.get("/api/v1/users/init/").asString())
-                .describedAs("Super user have to be created")
-                .isEqualTo("ADMIN USER CREATED");
-
-        token = flagSmithAuthenticate();
-        organisationId = createOrganisation("Test Organisation");
-    }
-
-    @AfterClass
-    public static void afterClass() {
-        backend.stop();
-        postgres.stop();
-        network.close();
-    }
 
     @Test(groups = "integration")
     public void testClient_When_Get_Features_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_Features_Then_Success",
                 "TEST");
 
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "Flag 1",
                 "Description for Flag 1",
                 environment.projectId,
                 true));
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "Flag 2",
                 "Description for Flag 2",
                 environment.projectId,
                 false));
-        createFeature(new ConfigFeature(
+        createFeature(new FlagsmithTestHelper.ConfigFeature(
                 "Config 1",
                 "Description for Config 1",
                 environment.projectId,
                 "xxx"));
-        createFeature(new ConfigFeature(
+        createFeature(new FlagsmithTestHelper.ConfigFeature(
                 "Config 2",
                 "Description for Config 2",
                 environment.projectId,
@@ -128,11 +68,11 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Has_Feature_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Has_Feature_Then_Success",
                 "TEST");
 
-        final int featureId = createFeature(new FlagFeature(
+        final int featureId = createFeature(new FlagsmithTestHelper.FlagFeature(
                 "Flag disabled",
                 null,
                 environment.projectId,
@@ -148,16 +88,16 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_Features_For_User_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_Features_For_User_Then_Success",
                 "TEST");
 
-        final int featureId = createFeature(new FlagFeature(
+        final int featureId = createFeature(new FlagsmithTestHelper.FlagFeature(
                 "Flag to be enabled for the user",
                 null,
                 environment.projectId,
                 false));
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "Other Flag",
                 null,
                 environment.projectId,
@@ -197,7 +137,7 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Traits_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Traits_Then_Success",
                 "TEST");
 
@@ -221,7 +161,7 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Traits_For_Keys_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Traits_For_Keys_Then_Success",
                 "TEST");
 
@@ -245,7 +185,7 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Traits_For_Invalid_User_Then_Return_Empty() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Traits_For_Invalid_User_Then_Return_Empty",
                 "TEST");
 
@@ -260,16 +200,16 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Traits_And_Flags_For_Keys_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Traits_And_Flags_For_Keys_Then_Success",
                 "TEST");
 
-        final int featureId = createFeature(new FlagFeature(
+        final int featureId = createFeature(new FlagsmithTestHelper.FlagFeature(
                 "Flag to be enabled for the user",
                 null,
                 environment.projectId,
                 false));
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "Other Flag",
                 null,
                 environment.projectId,
@@ -301,11 +241,11 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Traits_And_Flags_For_Invalid_User_Then_Return_Empty() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Traits_And_Flags_For_Invalid_User_Then_Return_Empty",
                 "TEST");
 
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "The Flag",
                 null,
                 environment.projectId,
@@ -327,11 +267,11 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Trait_From_Traits_And_Flags_For_Keys_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Trait_From_Traits_And_Flags_For_Keys_Then_Success",
                 "TEST");
 
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "The Flag",
                 null,
                 environment.projectId,
@@ -352,11 +292,11 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Traits_From_Traits_And_Flags_For_Keys_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Traits_From_Traits_And_Flags_For_Keys_Then_Success",
                 "TEST");
 
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "The Flag",
                 null,
                 environment.projectId,
@@ -383,16 +323,16 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_FLag_Value_From_Traits_And_Flags_For_Keys_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_FLag_Value_From_Traits_And_Flags_For_Keys_Then_Success",
                 "TEST");
 
-        createFeature(new FlagFeature(
+        createFeature(new FlagsmithTestHelper.FlagFeature(
                 "The Flag",
                 null,
                 environment.projectId,
                 false));
-        createFeature(new ConfigFeature(
+        createFeature(new FlagsmithTestHelper.ConfigFeature(
                 "font_size",
                 null,
                 environment.projectId,
@@ -410,11 +350,11 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_FLag_Enabled_From_Traits_And_Flags_For_Keys_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_FLag_Enabled_From_Traits_And_Flags_For_Keys_Then_Success",
                 "TEST");
 
-        final int featureId = createFeature(new FlagFeature(
+        final int featureId = createFeature(new FlagsmithTestHelper.FlagFeature(
                 "The Flag",
                 null,
                 environment.projectId,
@@ -434,7 +374,7 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Trait_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Trait_Then_Success",
                 "TEST");
 
@@ -448,7 +388,7 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Get_User_Trait_Update_Then_Updated() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Get_User_Trait_Update_Then_Updated",
                 "TEST");
 
@@ -470,7 +410,7 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Add_Traits_For_Identity_With_Missing_Identity_Then_Failed() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Add_Traits_For_Identity_With_Missing_Identity_Then_Failed",
                 "TEST");
 
@@ -483,7 +423,7 @@ public class FlagsmithClientTest {
 
     @Test(groups = "integration")
     public void testClient_When_Add_Traits_For_Identity_Then_Success() {
-        final ProjectEnvironment environment = createProjectEnvironment(
+        final FlagsmithTestHelper.ProjectEnvironment environment = createProjectEnvironment(
                 "testClient_When_Add_Traits_For_Identity_Then_Success",
                 "TEST");
 
@@ -499,316 +439,5 @@ public class FlagsmithClientTest {
                         trait(null, "trait_1", "some value1"),
                         trait(null, "trait_2", "some value2")
                 );
-    }
-
-    private ProjectEnvironment createProjectEnvironment(String projectName, String environmentName) {
-        final int projectId = createProject(projectName, organisationId);
-
-        final Map<String, Object> environment = createEnvironment(environmentName, projectId);
-        final String environmentApiKey = (String) environment.get("api_key");
-
-        final FlagsmithClient client = FlagsmithClient.newBuilder()
-                .withApiUrl(String.format("http://localhost:%d/api/v1/", backend.getMappedPort(BACKEND_PORT)))
-                .setApiKey(environmentApiKey)
-                .build();
-        return new ProjectEnvironment(environmentApiKey, projectId, client);
-    }
-
-    private int createFeature(Feature feature) {
-        final ImmutableMap.Builder<Object, Object> builder = ImmutableMap.builder()
-                .put("name", feature.name)
-                .put("type", feature.type)
-                .put("project", feature.projectId);
-        if (feature.description != null) {
-            builder.put("description", feature.description);
-        }
-
-        if (feature instanceof FlagFeature) {
-            builder.put("default_enabled", ((FlagFeature) feature).enabled);
-        } else if (feature instanceof ConfigFeature) {
-            builder.put("initial_value", ((ConfigFeature) feature).value);
-        } else {
-            throw new IllegalStateException("Unsupported feature " + feature);
-        }
-
-        return RestAssured.given()
-                .body(builder.build())
-                .headers(defaultHeaders())
-                .post("/api/v1/projects/{projectPk}/features/", feature.projectId)
-                .then()
-                .statusCode(201)
-                .extract()
-                .jsonPath()
-                .getInt("id");
-    }
-
-    private int createUserIdentity(String userIdentity, String environmentApiKey) {
-        return RestAssured.given()
-                .body(ImmutableMap.of(
-                        "identifier", userIdentity,
-                        "environment", environmentApiKey
-                ))
-                .headers(defaultHeaders())
-                .post("/api/v1/environments/{apiKey}/identities/", environmentApiKey)
-                .then()
-                .statusCode(201)
-                .extract()
-                .jsonPath()
-                .getInt("id");
-    }
-
-    private Map<String, Object> createEnvironment(String name, int projectId) {
-        return RestAssured.given()
-                .body(ImmutableMap.of(
-                        "name", name,
-                        "project", projectId
-                ))
-                .headers(defaultHeaders())
-                .post("/api/v1/environments/")
-                .then()
-                .statusCode(201)
-                .extract()
-                .jsonPath()
-                .getJsonObject("$");
-    }
-
-    private void switchFlag(int featureId, boolean enabled, String apiKey) {
-        final List<Map<String, Object>> features = RestAssured.given()
-                .headers(defaultHeaders())
-                .get("/api/v1/environments/{apiKey}/featurestates/?feature={featureId}",
-                        apiKey, featureId)
-                .then()
-                .statusCode(200)
-                .extract()
-                .jsonPath()
-                .getList("results");
-        if (features.isEmpty()) {
-            RestAssured.given()
-                    .body(ImmutableMap.of(
-                            "enabled", enabled,
-                            "feature", featureId
-                    ))
-                    .headers(defaultHeaders())
-                    .post("/api/v1/environments/{apiKey}/featurestates/", apiKey)
-                    .then()
-                    .statusCode(201);
-        } else {
-            final List<Integer> featureStateIds = features.stream()
-                    .map(jsonMap -> jsonMap.get("id"))
-                    .filter(Integer.class::isInstance)
-                    .map(Integer.class::cast)
-                    .collect(Collectors.toList());
-            if (featureStateIds.size() == 1) {
-                final int featureStateId = featureStateIds.get(0);
-                RestAssured.given()
-                        .body(ImmutableMap.of(
-                                "enabled", enabled,
-                                "feature", featureId
-                        ))
-                        .headers(defaultHeaders())
-                        .put("/api/v1/environments/{apiKey}/featurestates/{featureStateId}/",
-                                apiKey, featureStateId)
-                        .then()
-                        .statusCode(200);
-            } else {
-                throw new IllegalStateException("Unable to decide which feature-state to update: " + features);
-            }
-        }
-    }
-
-    private void switchFlagForUser(int featureId, int userIdentityId, boolean enabled, String apiKey) {
-        final List<Map<String, Object>> features = RestAssured.given()
-                .headers(defaultHeaders())
-                .get("/api/v1/environments/{apiKey}/identities/{identityId}/featurestates/?feature={featureId}",
-                        apiKey, userIdentityId, featureId)
-                .then()
-                .statusCode(200)
-                .extract()
-                .jsonPath()
-                .getList("results");
-        if (features.isEmpty()) {
-            RestAssured.given()
-                    .body(ImmutableMap.of(
-                            "enabled", enabled,
-                            "feature", featureId
-                    ))
-                    .headers(defaultHeaders())
-                    .post("/api/v1/environments/{apiKey}/identities/{identityId}/featurestates/", apiKey, userIdentityId)
-                    .then()
-                    .statusCode(201);
-        } else {
-            final List<Integer> featureStateIds = features.stream()
-                    .filter(jsonMap -> {
-                        final Map<String, Object> identityMap = (Map<String, Object>) jsonMap.get("identity");
-                        return Integer.valueOf(userIdentityId).equals(identityMap.get("id"));
-                    })
-                    .map(jsonMap -> jsonMap.get("id"))
-                    .filter(Integer.class::isInstance)
-                    .map(Integer.class::cast)
-                    .collect(Collectors.toList());
-            if (featureStateIds.size() == 1) {
-                final int featureStateId = featureStateIds.get(0);
-                RestAssured.given()
-                        .body(ImmutableMap.of(
-                                "enabled", enabled,
-                                "feature", featureId
-                        ))
-                        .headers(defaultHeaders())
-                        .put("/api/v1/environments/{apiKey}/identities/{identityId}/featurestates/{featureStateId}/",
-                                apiKey, userIdentityId, featureStateId)
-                        .then()
-                        .statusCode(200);
-            } else {
-                throw new IllegalStateException("Unable to decide which feature-state to update: " + features);
-            }
-        }
-    }
-
-    private void assignTraitToUserIdentity(String userIdentifier, String traitKey, Object traitValue, String apiKey) {
-        RestAssured.given()
-                .body(ImmutableMap.of(
-                        "identity", ImmutableMap.of("identifier", userIdentifier),
-                        "trait_key", traitKey,
-                        "trait_value", traitValue
-                ))
-                .headers(defaultHeaders())
-                .header("x-environment-key", apiKey)
-                .post("/api/v1/traits/")
-                .then()
-                .statusCode(200);
-    }
-
-    private int createProject(String name, int organisationId) {
-        return RestAssured.given()
-                .body(ImmutableMap.of(
-                        "name", name,
-                        "organisation", organisationId
-                ))
-                .headers(defaultHeaders())
-                .post("/api/v1/projects/")
-                .then()
-                .statusCode(201)
-                .extract()
-                .jsonPath()
-                .getInt("id");
-    }
-
-    private static int createOrganisation(String name) {
-        return RestAssured.given()
-                .body(ImmutableMap.of("name", name))
-                .headers(defaultHeaders())
-                .post("/api/v1/organisations/")
-                .then()
-                .statusCode(201)
-                .extract()
-                .jsonPath()
-                .getInt("id");
-    }
-
-    private static String flagSmithAuthenticate() {
-        return RestAssured.given()
-                .body(ImmutableMap.of(
-                        // from https://github.com/Flagsmith/flagsmith-api/blob/v2.5.0/src/app/settings/common.py#L255-L258
-                        "email", "admin@example.com",
-                        "password", "password"
-                ))
-                .header("Content-type", "application/json")
-                .post("/api/v1/auth/login/")
-                .then()
-                .statusCode(200)
-                .extract()
-                .jsonPath()
-                .getString("key");
-    }
-
-    private static Headers defaultHeaders() {
-        return new Headers(
-                new Header("Content-type", "application/json"),
-                new Header("Authorization", "Token " + token));
-    }
-
-    private Flag flag(String name, String description, boolean enabled) {
-        final com.flagsmith.Feature feature = new com.flagsmith.Feature();
-        feature.setName(name);
-        feature.setDescription(description);
-        feature.setType("FLAG");
-
-        final Flag result = new Flag();
-        result.setFeature(feature);
-        result.setEnabled(enabled);
-        return result;
-    }
-
-    private Flag config(String name, String description, String value) {
-        final com.flagsmith.Feature feature = new com.flagsmith.Feature();
-        feature.setName(name);
-        feature.setDescription(description);
-        feature.setType("CONFIG");
-
-        final Flag result = new Flag();
-        result.setFeature(feature);
-        result.setStateValue(value);
-        return result;
-    }
-
-    private Trait trait(String userIdentifier, String key, String value) {
-        final Trait result = new Trait();
-        if (userIdentifier != null) {
-            final FeatureUser user = featureUser(userIdentifier);
-            result.setIdentity(user);
-        }
-        result.setKey(key);
-        result.setValue(value);
-        return result;
-    }
-
-    private FeatureUser featureUser(String identifier) {
-        final FeatureUser user = new FeatureUser();
-        user.setIdentifier(identifier);
-        return user;
-    }
-
-    private static abstract class Feature {
-        final String type;
-        final String name;
-        final String description;
-        final int projectId;
-
-        protected Feature(String type, String name, String description, int projectId) {
-            this.type = type;
-            this.name = name;
-            this.description = description;
-            this.projectId = projectId;
-        }
-    }
-
-    private static class FlagFeature extends Feature {
-        final boolean enabled;
-
-        protected FlagFeature(String name, String description, int projectId, boolean enabled) {
-            super("FLAG", name, description, projectId);
-            this.enabled = enabled;
-        }
-    }
-
-    private static class ConfigFeature extends Feature {
-        final String value;
-
-        protected ConfigFeature(String name, String description, int projectId, String value) {
-            super("CONFIG", name, description, projectId);
-            this.value = value;
-        }
-    }
-
-    private static class ProjectEnvironment {
-        final String apiKey;
-        final int projectId;
-        final FlagsmithClient client;
-
-        private ProjectEnvironment(String apiKey, int projectId, FlagsmithClient client) {
-            this.apiKey = apiKey;
-            this.projectId = projectId;
-            this.client = client;
-        }
     }
 }
